When reading SMSC, ignore invalid SMSC
On Samsung, when reading SMSC we get weird strings like
"4392084000100",149
The first number is nowhere close to the SMSC phone number
This number is used when sending SMS, and since it is invalid, sending
SMS doesn't work.
We have no idea what this value means or anything, but let's just ignore
it. In 2024 we can send SMS just fine without a SMSC

This commit fixes sending SMS on OrangeF on my S24

Change-Id: I08c0a3cbb33f6a3a178e9a0d5b4df8f9fd7f669d